# openapi.yaml
openapi: 3.0.3
info:
  title: Zevaro Core API
  description: |
    Zevaro Core API for outcome-driven product development and decision management.

    ## Authentication
    This API uses JWT Bearer token authentication. Obtain tokens via the `/api/v1/auth/login` endpoint.

    ## Multi-tenancy
    All data is isolated by tenant. The tenant context is extracted from the JWT token.

    ## Rate Limiting
    Authentication endpoints are rate-limited. Exceeding limits returns HTTP 429.
  version: 1.0.0
  contact:
    name: Zevaro Support
    email: support@zevaro.ai
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.zevaro.ai
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/stakeholders:
    get:
      tags:
        - Stakeholders
      summary: List stakeholders
      description: Returns stakeholders with optional filtering.
      operationId: getStakeholders
      parameters:
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/StakeholderType'
        - name: activeOnly
          in: query
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: List of stakeholders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StakeholderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
        - Stakeholders
      summary: Create stakeholder
      description: Creates a new stakeholder.
      operationId: createStakeholder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStakeholderRequest'
      responses:
        '201':
          description: Stakeholder created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StakeholderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/stakeholders/{id}:
    get:
      tags:
        - Stakeholders
      summary: Get stakeholder by ID
      description: Returns a single stakeholder by their UUID.
      operationId: getStakeholder
      parameters:
        - $ref: '#/components/parameters/StakeholderId'
      responses:
        '200':
          description: Stakeholder details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StakeholderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Stakeholders
      summary: Update stakeholder
      description: Updates an existing stakeholder.
      operationId: updateStakeholder
      parameters:
        - $ref: '#/components/parameters/StakeholderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStakeholderRequest'
      responses:
        '200':
          description: Stakeholder updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StakeholderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Stakeholders
      summary: Delete stakeholder
      description: Deletes a stakeholder.
      operationId: deleteStakeholder
      parameters:
        - $ref: '#/components/parameters/StakeholderId'
      responses:
        '204':
          description: Stakeholder deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/stakeholders/email/{email}:
    get:
      tags:
        - Stakeholders
      summary: Get stakeholder by email
      description: Returns a stakeholder by their email address.
      operationId: getStakeholderByEmail
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: Stakeholder details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StakeholderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/stakeholders/with-pending:
    get:
      tags:
        - Stakeholders
      summary: Get stakeholders with pending decisions
      description: Returns stakeholders who have pending decisions awaiting input.
      operationId: getStakeholdersWithPending
      responses:
        '200':
          description: List of stakeholders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StakeholderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/stakeholders/slow-responders:
    get:
      tags:
        - Stakeholders
      summary: Get slow responders
      description: Returns stakeholders with above-threshold average response times.
      operationId: getSlowResponders
      parameters:
        - name: thresholdHours
          in: query
          schema:
            type: number
            format: double
            default: 24
      responses:
        '200':
          description: List of slow responders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StakeholderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/stakeholders/search/expertise:
    get:
      tags:
        - Stakeholders
      summary: Find by expertise
      description: Finds stakeholders with matching expertise.
      operationId: findByExpertise
      parameters:
        - name: expertise
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of stakeholders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StakeholderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v1/stakeholders/{id}/stats:
    get:
      tags:
        - Stakeholders
      summary: Get stakeholder stats
      description: Returns statistics for a stakeholder (alias for metrics).
      operationId: getStakeholderStats
      parameters:
        - $ref: '#/components/parameters/StakeholderId'
      responses:
        '200':
          description: Stakeholder statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StakeholderMetrics'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/stakeholders/{id}/responses:
    get:
      tags:
        - Stakeholders
      summary: Get stakeholder responses
      description: Returns paginated decision responses for a stakeholder.
      operationId: getStakeholderResponses
      parameters:
        - $ref: '#/components/parameters/StakeholderId'
        - name: page
          in: query
          description: Page number (0-indexed)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Page size
          schema:
            type: integer
            default: 20
        - name: pending
          in: query
          description: Filter by pending status
          schema:
            type: boolean
        - name: withinSla
          in: query
          description: Filter by SLA compliance
          schema:
            type: boolean
      responses:
        '200':
          description: Paginated list of decision responses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedDecisionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/stakeholders/leaderboard:
    get:
      tags:
        - Stakeholders
      summary: Get stakeholder leaderboard
      description: Returns stakeholder rankings by engagement metrics.
      operationId: getStakeholderLeaderboard
      responses:
        '200':
          description: Stakeholder leaderboard
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StakeholderLeaderboard'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/stakeholders/me/responses/pending:
    get:
      tags:
        - Stakeholders
      summary: Get my pending responses
      description: Returns decisions awaiting response from the current stakeholder.
      operationId: getMyPendingResponses
      responses:
        '200':
          description: List of pending decisions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DecisionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================
  # QUEUE ENDPOINTS
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/v1/auth/login
  parameters:
...
    StakeholderId:
      name: id
      in: path
      description: Stakeholder UUID
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        fullName:
          type: string
        avatarUrl:
          type: string
          format: uri
...
    TeamSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
...
    OutcomeSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        status:
          $ref: '#/components/schemas/OutcomeStatus'
...
    OutcomeStatus:
      type: string
      enum:
        - DRAFT
        - NOT_STARTED
        - IN_PROGRESS
        - VALIDATING
        - VALIDATED
        - INVALIDATED
        - ABANDONED
...
    DecisionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        context:
          type: string
        options:
          type: array
          items:
            $ref: '#/components/schemas/DecisionOption'
        status:
          $ref: '#/components/schemas/DecisionStatus'
        priority:
          $ref: '#/components/schemas/DecisionPriority'
        type:
          $ref: '#/components/schemas/DecisionType'
        owner:
          $ref: '#/components/schemas/UserSummary'
        assignedTo:
          $ref: '#/components/schemas/UserSummary'
        outcome:
          $ref: '#/components/schemas/OutcomeSummary'
        hypothesis:
          $ref: '#/components/schemas/HypothesisSummary'
        team:
          $ref: '#/components/schemas/TeamSummary'
        queue:
          $ref: '#/components/schemas/QueueSummary'
        stakeholder:
          $ref: '#/components/schemas/StakeholderSummary'
        slaHours:
          type: integer
        dueAt:
          type: string
          format: date-time
        overdue:
          type: boolean
        waitTimeHours:
          type: number
        escalationLevel:
          type: integer
        escalatedAt:
          type: string
          format: date-time
        escalatedTo:
          $ref: '#/components/schemas/UserSummary'
        decidedBy:
          $ref: '#/components/schemas/UserSummary'
        decidedAt:
          type: string
          format: date-time
        decisionRationale:
          type: string
        selectedOption:
          $ref: '#/components/schemas/DecisionOption'
        resolution:
          type: string
        wasEscalated:
          type: boolean
        blockedItems:
          type: array
          items:
            $ref: '#/components/schemas/BlockedItem'
        commentCount:
          type: integer
        voteCount:
          type: integer
        externalRefs:
          type: object
          additionalProperties:
            type: string
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DecisionOption:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        pros:
          type: array
          items:
            type: string
        cons:
          type: array
          items:
            type: string
    BlockedItem:
      type: object
      properties:
        type:
          type: string
        id:
          type: string
        title:
          type: string
        url:
          type: string
...
    DecisionStatus:
      type: string
      enum:
        - NEEDS_INPUT
        - UNDER_DISCUSSION
        - DECIDED
        - IMPLEMENTED
        - DEFERRED
        - CANCELLED
    DecisionPriority:
      type: string
      enum:
        - BLOCKING
        - HIGH
        - NORMAL
        - LOW
      description: |
        Priority levels with default SLA hours:
        - BLOCKING: 4 hours
        - HIGH: 8 hours
        - NORMAL: 24 hours
        - LOW: 72 hours
    DecisionType:
      type: string
      enum:
        - PRODUCT
        - UX
        - TECHNICAL
        - ARCHITECTURAL
        - STRATEGIC
        - OPERATIONAL
        - RESOURCE
        - SCOPE
        - TIMELINE
...
    PagedDecisionResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/DecisionResponse'
        pageable:
          $ref: '#/components/schemas/Pageable'
        totalPages:
          type: integer
        totalElements:
          type: integer
          format: int64
        last:
          type: boolean
        first:
          type: boolean
        size:
          type: integer
        number:
          type: integer
        numberOfElements:
          type: integer
        empty:
          type: boolean

    # ============================================
    # COMMENT & VOTE SCHEMAS
    # ============================================
...
    HypothesisSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        status:
          $ref: '#/components/schemas/HypothesisStatus'
...
    HypothesisStatus:
      type: string
      enum:
        - DRAFT
        - READY
        - BLOCKED
        - BUILDING
        - DEPLOYED
        - MEASURING
        - VALIDATED
        - INVALIDATED
        - ABANDONED
...
    StakeholderResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        title:
          type: string
        organization:
          type: string
        phone:
          type: string
        avatarUrl:
          type: string
          format: uri
        type:
          $ref: '#/components/schemas/StakeholderType'
        expertise:
          type: string
        preferredContactMethod:
          type: string
        availabilityNotes:
          type: string
        timezone:
          type: string
        decisionsPending:
          type: integer
        decisionsCompleted:
          type: integer
        decisionsEscalated:
          type: integer
        avgResponseTimeHours:
          type: number
        lastDecisionAt:
          type: string
          format: date-time
        active:
          type: boolean
        notes:
          type: string
        externalRefs:
          type: object
          additionalProperties:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    StakeholderSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        type:
          $ref: '#/components/schemas/StakeholderType'
    CreateStakeholderRequest:
      type: object
      required:
        - name
        - email
        - type
      properties:
        name:
          type: string
          maxLength: 200
        email:
          type: string
          format: email
        type:
          $ref: '#/components/schemas/StakeholderType'
        title:
          type: string
          maxLength: 200
        organization:
          type: string
          maxLength: 200
        phone:
          type: string
          maxLength: 50
        expertise:
          type: string
          maxLength: 500
        preferredContactMethod:
          type: string
          maxLength: 100
        timezone:
          type: string
          maxLength: 50
        notes:
          type: string
          maxLength: 2000
    UpdateStakeholderRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
        email:
          type: string
          format: email
        type:
          $ref: '#/components/schemas/StakeholderType'
        title:
          type: string
          maxLength: 200
        organization:
          type: string
          maxLength: 200
        phone:
          type: string
          maxLength: 50
        expertise:
          type: string
          maxLength: 500
        preferredContactMethod:
          type: string
          maxLength: 100
        timezone:
          type: string
          maxLength: 50
        notes:
          type: string
          maxLength: 2000
        active:
          type: boolean
    StakeholderType:
      type: string
      enum:
        - INTERNAL
        - EXTERNAL
        - EXECUTIVE
        - CUSTOMER
        - REGULATORY
        - TECHNICAL
    StakeholderMetrics:
      type: object
      properties:
        stakeholderId:
          type: string
          format: uuid
        totalDecisions:
          type: integer
        pendingDecisions:
          type: integer
        completedDecisions:
          type: integer
        escalatedDecisions:
          type: integer
        avgResponseTimeHours:
          type: number
        responseTimeP50Hours:
          type: number
        responseTimeP90Hours:
          type: number
    StakeholderLeaderboard:
      type: object
      properties:
        fastestResponders:
          type: array
          items:
            $ref: '#/components/schemas/StakeholderResponse'
        mostActive:
          type: array
          items:
            $ref: '#/components/schemas/StakeholderResponse'
        needsAttention:
          type: array
          items:
            $ref: '#/components/schemas/StakeholderResponse'

    # ============================================
    # QUEUE SCHEMAS
    # ============================================
...
    QueueSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
          description: Error code
          enum:
            - NOT_FOUND
            - FORBIDDEN
            - UNAUTHORIZED
            - BAD_REQUEST
            - CONFLICT
            - VALIDATION_ERROR
            - INTERNAL_ERROR
        message:
          type: string
          description: Human-readable error message
        errors:
          type: object
          additionalProperties:
            type: string
          description: Field-level validation errors
    Pageable:
      type: object
      properties:
        pageNumber:
          type: integer
        pageSize:
          type: integer
        sort:
          type: array
          items:
            type: object
            properties:
              property:
                type: string
              direction:
                type: string
        offset:
          type: integer
          format: int64
        paged:
          type: boolean
        unpaged:
          type: boolean